Edit Customer Type
