+++ ACSS/data/customerFeedbacks.txt
@@ -4,3 +4,4 @@
 14cb8732,S004,SALESMAN,2,Rude
 14cb8732,S004,SALESMAN,2,Rude
 14cb8732,S004,SALESMAN,2,Rude
+14cb8731,C001,CAR_PURCHASED,5,good
